Fix regression in bubbleSortVersion
...introduced with a7b525d07112aaa4c347989c6c0e70a50e727226 "clang-tidy
performance-unnecessary-copy-initialization" (so partially revert it).  Whatever
clang-tidy erroneously reported there, cur and next are lvalue references into
vec, so this attempted copy now actually overwrote one with the other.  The
result was that if multiple JREs are detected on the system, "Options -
LibreOffice - Advanced" would list a single one multiple times.

Change-Id: I7ef454c0f37669722812383848602dc2bacf7cd1