drawing shapes aren't tracking style name changes correctly
in draw, create a box, enter some text, exit and select box
F11 for style navigator, new style, set font to e.g. 88
apply this style to the box. All ok.

If you modify the style and changes its font to e.g. bold
the text in the box tracks it as expected. All ok.

If you modify the style and change its name, the box resets
to something different and there's loads of warnings about
unknown styles. Not ok.

The style modification fires, but the style name gets overwritten
by the style stuck in a OutlinerParaObject that gets overlaid
over the editengine.

All the rTextProvider.getTextCount uses that don't iterate over all entries
look dubious to me, but I'll initially fix this one that I have hard evidence
for as broken.

All this is probably broken since...

commit 838c0fa8228f4ca656a264f6a5610c337ebf4fef
Author: Rüdiger Timm <rt@openoffice.org>
Date:   Wed Mar 12 08:47:30 2008 +0000

    INTEGRATION: CWS impresstables2 (1.14.72); FILE MERGED

Change-Id: Ib0dd2857ed26c49dcfa00aae923e36429a7f7862