configure: fix harfbuzz build on WNT
The graphite library is always built static, so the graphite headers must
be included with the GRAPHITE2_STATIC macro, so that the functions are
not marked __declspec(dllimport) on WNT.

Add GRAPHITE2_STATIC to GRAPHITE_CFLAGS so harfbuzz picks it up.

It's a mystery why vcl even links currently without this.

Change-Id: I80d87fe76afa2d05a5697d56dfaf8cbe97d9f977