prevent endlessly creating new lines until out of memory, tdf#104152
Though this just prevents mishappens during reformatting a broken portion that
is in a half-ready state because a11y was notified and attempts to obtain a
paragraph. Solving the real cause means not being called in such state..
However, not going out of memory and also having matching X-values for the
node's text positions that don't make the EditEngine crash later is better than
nothing.

Change-Id: Ifa8d2c5d1766b0f145b8e89ad26a78528db42056