Drop font cache on UNX
On my system reading the ~400 fonts I have takes ~0.07 seconds, which
does not really justify all this complexity. We don’t have such cache on
other platforms as well.

It might have been slower in the past with all PFB and AFM parsing, but
this is gone already.

Killing this ugly fontmanager over engineering one piece at a time.

Change-Id: I41fe3db48dc3de0cf8939c2120403f7d243d6096
Reviewed-on: https://gerrit.libreoffice.org/31511
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Khaled Hosny <khaledhosny@eglug.org>