Require HarfBuzz 0.9.42
This is the first version where hb_buffer_set_cluster_level() was
introduced. The layout engine depends on this feature in a few places
and though we can build without it, the result will be wrong in some
cases (i.e. any where we assume combining marks will have their own
cluster number).

Change-Id: I89187f317052e7933b83ad0350cace239333510b
Reviewed-on: https://gerrit.libreoffice.org/31521
Reviewed-by: Rene Engelhard <rene@debian.org>
Reviewed-by: Khaled Hosny <khaledhosny@eglug.org>
Tested-by: Khaled Hosny <khaledhosny@eglug.org>