tdf#104434 tdf#104451 Revert "tdf#46738 Fix exporting .xlsx of coloured em"...
This reverts commit 576f7bd6c929eb3480af03f3243791b3583ea65c.

That lead to including / looping through all formatted rows, which
easily happens if an entire column is formatted.

Also when printing it would print all formatted rows and printing a
million empty coloured rows certainly is not desired, only one sheet of
paper maybe. Which was the original intention of SC_VISATTR_STOP and as
the name GetLastVisibleAttr() suggests the function is for visible
attribution.