Resolves: tdf#104097 register reserved cell style names for auto styles
SvXMLAutoStylePoolP_Impl re-starts with a clean
XMLAutoStyleFamily::maReservedNameSet, so reserving the names in
ScXMLExport::ExportStyles_() is not enough and needs to be redone in
ScXMLExport::ExportAutoStyles_()

In the bug scenario that lead to cell styles with duplicated names for
different styles of which when reloading only the first was taken.

Additionally, a style must be registered after SvXMLAutoStylePoolP::AddNamed()
to prevent duplicated styles with different names.

Change-Id: If3491eb89ab1741887d9a36ecf904a216fb1368e