tdf#104046 - Slides in the slide pane don't update in realtime
This appears to be a consequence of my change

    commit 4863ce021f051673569c7d11d34a676910f3a78b
    convert EID constants to typed_flags

in that change I made a "fix":
    @@ -689,20 +687,20 @@ void
EventMultiplexer::Implementation::CallListeners (EventMultiplexerEvent&
rEv
         ListenerList::const_iterator iListenerEnd
(aCopyListeners.end());
         for (; iListener!=iListenerEnd; ++iListener)
         {
    -        if ((iListener->second && rEvent.meEventId))
    +        if (iListener->second & rEvent.meEventId)
             iListener->first.Call(rEvent);
         }
     }

which causes this bug.

I should have noticed that my "fix" indicates that the event filtering
part of this multiplexing code was never working, and since no-one has
ever complained about, lets just remove all of this unnecessary
complexity.

Change-Id: Id71613d4fd5817ee1358705059e4ce63d57573ad
Reviewed-on: https://gerrit.libreoffice.org/31894
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>