Check that VclPtr is only instantiated with appropriate types
...deriving from VclReferenceBase.  Complicated by the fact that the argument
type may be incomplete at the time of template instantiation.  So this approach
may be less precise than the change to loplugin:vclwidgets from
7602eb5b229df219c13a30ae14605d2d42f47512 "Catch some misuses of VclPtr
construction" when the argument type becomes complete later in the comilation
unit.  However, this approach would also catch the two misuses in UnoControls
found by 7602eb5b229df219c13a30ae14605d2d42f47512, so go with this approach for
now and revert the change to loplugin:vclwdigets.

Change-Id: I7888f23d2b9e2db81ae2ce4bf4c8277912317685
Reviewed-on: https://gerrit.libreoffice.org/31966
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Stephan Bergmann <sbergman@redhat.com>