tdf#91260 sw: fix nasty corner case of SdrGroupObject in header
or footer.  Some very special handling going on to paint these in
VOCOfDrawVirtObj with a GetSwDrawVirtObj().GetOffset() added;
they are falsely identified as invisible because their
sub-objects have an anchor set but it is always on the first page
so to properly check them we'd have to add the same GetOffset()
thing but checking the top-level object should be sufficient.

(regression from c444113249df756c698e8b145eaf78259d5a0bab)

Change-Id: Ida6fcd41948f46410ca02193549e426bc9dae8c5