tdf#104488 SwAccessibleMap: dispose sub-shapes of group shapes
... and ensure that they are removed from the mpShapeMap.

Commit 15a1104df509af780cbe45c6e3e299afe9b8606c added all sub-shapes of
group shapes to SwAccessibleMap::mpShapeMap.
It's not immediately obvious why this is necessary, given that
AccessibleShape will probably create these on demand.

It turns out that the AccessibleShapes are not only disposed from
SwAccessible code but also from svx code,
in particular the call to ViewForwarderChanged() in
SwAccessibleContext::ChildrenScrolled() leads to visibility checks.

Thus the RemoveGroupContext() may fail to remove the sub-shapes from the
mpShapeMap and the assertion in ~SwAccessibleMap is triggered.

Change-Id: I6d9c3d816f5521960855d5e6a563bec49d8030b8