Take advantage of Unicode, people
As a translator, I’ve always found a bit ridiculous that LibreOffice
provides me with separate strings for opening and closing quotation
marks, but the English version doesn’t actually take advantage of it.
It seems that, because of this, many translators haven’t taken advantage
of the separate strings. Let’s fix that.

Change-Id: I125fcf311916dfdb056e5a46e081fa9d0305db25