tdf#104492: Fix bad docx opening and bad layout
The problem seems to be the table's repeated rows value which is
setted to max table rows in some docx file.

Decreasing repeated rows value by one avoids loops but renders a bad
layout.
Table continuity is lost and lots of rows are hidden.

So I set repeated rows value directly to one in order to avoid bad
layout.

Change-Id: If4adc33febe088a866b704e87e49339710e5aaf5
Reviewed-on: https://gerrit.libreoffice.org/32349
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Björn Michaelsen <bjoern.michaelsen@canonical.com>