tdf#97228 Moved class ControlCacheKey in a seperate include file.
class ControlCacheKey defined in include/vcl/salnativewidgets.hxx is used
only by vcl therefore it is moved to a seperate include file named
ControlCacheKey.hxx which is inside vcl/inc/ControlCacheKey.hxx.

This header file contains the definition of the the class ControlCacheKey.

Since this class uses an object of another class named Size which is
defined in gen.hxx therefore this is included in ControlCacheKey.hxx.

All the files that references the class ControlCacheKey includes
ControlCacheKey.hxx with the syntax: #include "ControlCacheKey.hxx"

Change-Id: I7d7b3e1cf6950a953e6bbed670c1914826b2bacd
Reviewed-on: https://gerrit.libreoffice.org/32422
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Michael Stahl <mstahl@redhat.com>