Move documentation for parameter of SfxObjectShell::GetTitle() to include file
A re-design of this API would be a good idea.

Is this function ever called with some specific max length that would
not be one of the magic values? Probably the sanest would be to turn
the magic value constants into a normal (scoped) enum, and have two
separate functions: One that takes no parameter and returns the whole
title (and then let the caller take care of truncating as necessary),
and one that takes such an enum and returns a correspondingly mangled
title. The less magic and defaulted parameters the better.

Change-Id: I9e3aa95fae981c5ed48300b753556fa9074c1166